(cleanup) moved query to queries.js
